requirements.txt
fastapi
uvicorn
torch
timm
pillow
torchvision
python-multipart
